Added cows!
Cows now exist, and they give you propellant! How wonderful! How smelly!
Also put in some stuff for a possible particle system (for exhaust and such), but this will likely not be feasible for performance reasons (this is not particularly efficient, and even a 256x256 cow image slows down the game significantly)